Add gamedev page
For now I just wanted to hastily write down the links so I don't lose
them. This commit should be followed up with making them proper
hyperlinks.

I am contemplating how this should be done. Writing the title of the
video and making that a link is obviously very clean but it has the
arguable downside of hiding the domain. On the other hand, raw links can
be very verbose and ugly. Maybe the best solution is to note the domain,
eg

    Title of Talk [youtube.com]

or some equivalent, where maybe only the [...] part is clickable to make
it very obvious that the user understand where they are being
redirected.